requirements.txt
feedparser>=6.0.10
requests>=2.31.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.0
openai>=1.0.0
anthropic>=0.18.0
google-genai>=0.4.0
